+++ lib/src/utils/widgets.dart
@@ -9,7 +9,7 @@ export 'repeat.dart';
 export 'foreach.dart';
 export 'do_loop.dart';
 export 'raycast.dart';
-export 'builder.dart';
+export '../basic/builder.dart';
 export 'storage.dart';
 export 'server_version_check.dart';
 export 'version_check.dart';